misc/22_secret-server/src/requirements.txt
discord.py==2.4.0
python-dotenv
SQLAlchemy
